+++ src/card/model/presentation.rs
@@ -80,6 +80,7 @@ fn object_predicate_implies(predicate: ObjectPredicateDef, expected: ObjectPredi
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -125,6 +126,7 @@ fn predicate_color(predicate: ObjectPredicateDef) -> Option<ManaColor> {
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -172,6 +174,7 @@ fn predicate_color_count(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -219,6 +222,7 @@ fn predicate_subtype(predicate: ObjectPredicateDef) -> Option<&'static str> {
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -270,6 +274,7 @@ fn predicate_negated_subtype(predicate: ObjectPredicateDef) -> Option<&'static s
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -326,6 +331,7 @@ fn predicate_power_at_least(predicate: ObjectPredicateDef) -> Option<i16> {
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -368,6 +374,7 @@ fn predicate_mana_value_at_most(predicate: ObjectPredicateDef) -> Option<u8> {
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -415,6 +422,7 @@ fn predicate_controller(predicate: ObjectPredicateDef) -> Option<PlayerRelation>
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource
@@ -465,6 +473,7 @@ fn predicate_negates(predicate: ObjectPredicateDef, expected: ObjectPredicateDef
         | ObjectPredicateDef::Attacking
         | ObjectPredicateDef::Blocking
         | ObjectPredicateDef::BlockedBySource
+        | ObjectPredicateDef::BlockingSource
         | ObjectPredicateDef::Enchanted
         | ObjectPredicateDef::AttachedTo(_)
         | ObjectPredicateDef::AttachedToSource